The 923 nodes with no box are not bounds waiting to be written
Step 2 of the order asked for the histogram before any bounds, and the
histogram refutes the first half of the step.

Field::unbounded_by_op splits every unbounded node two ways: whether its
own case in build_bounds gave up or a child took its box away, and
whether the solid can reach it at all. --clip-field reaches the table
without sampling anything, because the parse decides all of it.

On the facility: 923 of 3,744 nodes have no box, but only 177 are under
the solid, and a shape evaluation walks nothing else. Of those 177, 87
are non-uniform scales refused on purpose, 59 are half spaces that are
genuinely infinite, and the rest are values with no extent. 54 nodes of
the 923 are unbounded by inheritance, against the 38% the rotate case
was written to cure. There is no missing case to write.

What is left of the step is kAccelerateFrom - 190 wide unions have no
hierarchy and that constant has never been measured. D636 records it,
along with a sound cull for non-uniform scales that is deliberately not
being built until that arm says whether these unions cull at all.

The tree also builds and tests on Linux again, which is how this was
measured: gcc's unused-function and enum-conditional warnings are errors
there, an unread distance_to helper is gone, and Vulkan's designated
initialisers stop being several hundred missing-field errors. Two
library tests are skipped off Windows because send_to_recycle_bin has no
Linux body - a real gap, tracked as debt.

536 tests, 2 skipped, 18.7 M assertions. clips/sampler.clip still
measures 1,430,104 voxels.

+++ src/world/node_pool.cpp
@@ -1260,7 +1260,8 @@ const NodeUploadBatch& NodePool::update(const World& world, const f64 camera_vox
         const NodeKey& key = requested_[asked];
         if (!seen_requests_.insert(key).second) continue;
         const u8 source =
-            (asked < requested_source_.size()) ? requested_source_[asked] : kRequestRay;
+            (asked < requested_source_.size()) ? requested_source_[asked]
+                                               : static_cast<u8>(kRequestRay);
 
         // Instrument: is this a node this pool threw away a moment ago?
         //
